master

+++ views/reserva/mostrarreserves.php
@@ -1,6 +1,8 @@
 <table class="table">
 <tr>
 <th>Codi vol</th>
+<th>Origen</th>
+<th>Desti</th>
 <th>Data Anada</th>
 <th>Data tornada</th>
 <th>Nombre de places</th>
@@ -9,16 +11,22 @@
 <th>Pagar ticket</th>
 </tr>
 <?php
+@session_start();
 while($row = $reserves->fetch_assoc()){
-    echo "<tr>";
-    echo "<td>".$row['codi_vol']."</td>";
-    echo "<td>".$row['data_anada']."</td>";
-    echo "<td>".$row['data_tornada']."</td>";
-    echo "<td>".$row['nombre_places']."</td>";
-    echo "<td><a href='index.php?controller=reserva&action=modificarreserves&codi=".$row['codi']."'>Modificar</a></td>";
-    echo "<td><a href='index.php?controller=reserva&action=eliminarreserves&codi=".$row['codi']."'>Eliminar</a></td>";
-    echo "<td><a href='index.php?controller=ticket&action=pagartickets&codi=".$row['codi']."'>Pagar</a></td>";
-    echo "</tr>";
+    if($row['codi_usuari'] == $_SESSION['codi'] || $_SESSION['rol'] == 'admin'){
+        echo "<tr>";
+        echo "<td>".$row['codi_vol']."</td>";
+        echo "<td>".$row['origen']."</td>";
+        echo "<td>".$row['desti']."</td>";
+        echo "<td>".$row['data_anada']."</td>";
+        echo "<td>".$row['data_tornada']."</td>";
+        echo "<td>".$row['nombre_places']."</td>";
+        echo "<td><a href='index.php?controller=reserva&action=modificarreserves&codi=".$row['codi']."'>Modificar</a></td>";
+        echo "<td><a href='index.php?controller=reserva&action=eliminarreserves&codi=".$row['codi']."'>Eliminar</a></td>";
+        echo "<td><a href='index.php?controller=ticket&action=pagartickets&codi=".$row['codi']."'>Pagar</a></td>";
+        echo "</tr>";
+    }
+
 }
 ?>
 </table>
